+++ packages/generator/src/lib/adapter/modules/createModelModule.ts
@@ -1,8 +1,8 @@
 import { DMMF } from '@prisma/generator-helper'
-import { generateTableDeclaration } from '../../adapter/declarations/generateTableDeclaration'
 import { Context } from '../../context'
 import { getModelModuleName } from '../../prisma-helpers/model'
-import { createModule } from './module'
+import { createModule } from '../../syntaxes/module'
+import { generateTableDeclaration } from '../declarations/generateTableDeclaration'
 
 export function createModelModule(input: { model: DMMF.Model; ctx: Context }) {
   const tableVar = generateTableDeclaration(input.ctx.adapter, input.model)
